Made Reg FSM less messy
Reg FSM now doesn't need a SlowClk. Everything is inside one always
block that runs off posedge of the Clk. It now looks after the 16th
register, and the reset button works now too. Also made a small change
to the RegisterFile to make it neater and actually implement a mux.